+++ crates/client/mempool/src/lib.rs
@@ -1,19 +1,16 @@
 use blockifier::blockifier::stateful_validator::StatefulValidatorError;
 use blockifier::transaction::account_transaction::AccountTransaction;
 use blockifier::transaction::transaction_execution::Transaction;
-use blockifier::transaction::transactions::DeployAccountTransaction;
-use blockifier::transaction::transactions::InvokeTransaction;
-use blockifier::transaction::transactions::{DeclareTransaction, L1HandlerTransaction};
+use blockifier::transaction::transactions::{
+    DeclareTransaction, DeployAccountTransaction, InvokeTransaction, L1HandlerTransaction,
+};
 use header::make_pending_header;
 use inner::MempoolInner;
 use mc_db::db_block_id::DbBlockId;
-use mc_db::MadaraBackend;
-use mc_db::MadaraStorageError;
+use mc_db::{MadaraBackend, MadaraStorageError};
 use mc_exec::ExecutionContext;
 use metrics::MempoolMetrics;
-use mp_block::BlockId;
-use mp_block::BlockTag;
-use mp_block::MadaraPendingBlockInfo;
+use mp_block::{BlockId, BlockTag, MadaraPendingBlockInfo};
 use mp_class::ConvertedClass;
 use mp_convert::ToFelt;
 use mp_transactions::{
@@ -22,16 +19,12 @@ use mp_transactions::{
 use mp_transactions::{BroadcastedToBlockifierError, L1HandlerTransactionResult};
 use starknet_api::core::{ContractAddress, Nonce};
 use starknet_api::transaction::TransactionHash;
-use starknet_core::types::BroadcastedDeclareTransaction;
-use starknet_core::types::BroadcastedDeployAccountTransaction;
-use starknet_core::types::BroadcastedInvokeTransaction;
-use starknet_core::types::BroadcastedTransaction;
-use starknet_core::types::DeclareTransactionResult;
-use starknet_core::types::DeployAccountTransactionResult;
-use starknet_core::types::InvokeTransactionResult;
+use starknet_core::types::{
+    BroadcastedDeclareTransaction, BroadcastedDeployAccountTransaction, BroadcastedInvokeTransaction,
+    BroadcastedTransaction, DeclareTransactionResult, DeployAccountTransactionResult, InvokeTransactionResult,
+};
 use starknet_types_core::felt::Felt;
-use std::sync::Arc;
-use std::sync::RwLock;
+use std::sync::{Arc, RwLock};
 
 pub use inner::TxInsersionError;
 pub use inner::{ArrivedAtTimestamp, MempoolTransaction};
@@ -359,7 +352,7 @@ pub(crate) fn clone_transaction(tx: &Transaction) -> Transaction {
 mod test {
     use std::sync::Arc;
 
-    use starknet_core::types::Felt;
+    use starknet_types_core::felt::Felt;
 
     use crate::MockL1DataProvider;