+++ crates/client/rpc/src/versions/v0_7_1/methods/read/get_block_with_txs.rs
@@ -1,8 +1,6 @@
-use starknet_core::types::{BlockId, MaybePendingBlockWithTxs};
-
 use jsonrpsee::core::RpcResult;
-use mp_block::MadaraMaybePendingBlockInfo;
-use starknet_core::types::{BlockStatus, BlockWithTxs, PendingBlockWithTxs};
+use mp_block::{BlockId, MadaraMaybePendingBlockInfo};
+use starknet_core::types::{BlockStatus, BlockWithTxs, MaybePendingBlockWithTxs, PendingBlockWithTxs};
 
 use crate::Starknet;
 
@@ -75,8 +73,10 @@ mod tests {
         errors::StarknetRpcApiError,
         test_utils::{sample_chain_for_block_getters, SampleChainForBlockGetters},
     };
+    use mp_block::BlockTag;
     use rstest::rstest;
-    use starknet_core::types::{BlockTag, Felt, L1DataAvailabilityMode, ResourcePrice};
+    use starknet_core::types::{L1DataAvailabilityMode, ResourcePrice};
+    use starknet_types_core::felt::Felt;
 
     #[rstest]
     fn test_get_block_with_txs(sample_chain_for_block_getters: (SampleChainForBlockGetters, Starknet)) {