+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## Next release
 
-- build(fgw): added `add_transaction` for gateway client
+- refactor: replace starknet-rs BlockId with types-rs BlockId and remove redundant mp_block::BlockId
+- feat(fgw): added `add_transaction` for gateway client
 - fix(fgw): include `l1_to_l2_consumed_message` in L1 handler receipt
 - build: up starknet-rs, starknet-types, blockifier(v0.8.0), cairo
 - feat(rpc): added `getCompiledCasm` method